[backend] move indicator to module (#2859)
